move fake orthology data
